requirements.txt
PyQt5>=5.15.0
tensorflow>=2.0.0
keras>=2.3.0
Pillow>=8.0.0
PyYAML>=5.1
opencv-python>=4.5.0
numpy>=1.18.0
matplotlib>=3.1.0
scikit-learn>=0.22.0
imutils>=0.5.3
natsort>=7.1.0
